requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
psycopg2-binary
python-decouple
gunicorn
